+++ config.py
@@ -1,0 +1,4 @@
+import os
+
+ROOT_DIR = os.path.dirname(__file__)
+DATA_PATH = os.path.join(ROOT_DIR, 'data')
