+++ main.py
@@ -1,4 +1,8 @@
-from src.funcs import user_interaction
+from src.vacancy import Vacancy
+def main():
+    v = Vacancy()
+    print(v)
+
 
 if __name__ == "__main__":
     main()
