+++ src/saver.py
@@ -1,28 +1,7 @@
 from abc import ABC, abstractmethod
 
 
-class AbstractAPI(ABC):
-    """Создаем абстрактный класс для класса APIconnect"""
-
-    @abstractmethod
-    def __init__(self):
-        pass
-
-    @abstractmethod
-    def __str__(self):
-        pass
-
-    @abstractmethod
-    def __repr__(self):
-        pass
-
-    @abstractmethod
-    def get_vacancy(self):
-        pass
-
-
-
-class AbstractVacancy(ABC):
+class SaverJSON(ABC):
     """Создаем абстрактный класс для класса Vacancy"""
 
     @abstractmethod
